fix(chat): keep chat path links helper in server tree
